Fix day 15 performance issue

+++ src/y2020/day15/Input.java
@@ -1,12 +1,12 @@
 package y2020.day15;
 
 public interface Input {
-    long[] sample1 = new long[] { 0L, 3L, 6L };
-    long[] sample2 = new long[] { 1L, 3L, 2L };
-    long[] sample3 = new long[] { 2L, 1L, 3L };
-    long[] sample4 = new long[] { 1L, 2L, 3L };
-    long[] sample5 = new long[] { 2L, 3L, 1L };
-    long[] sample6 = new long[] { 3L, 2L, 1L };
-    long[] sample7 = new long[] { 3L, 1L, 2L };
-    long[] puzzle = new long[] { 9L, 6L, 0L, 10L, 18L, 2L, 1L };
+    int[] sample1 = new int[] { 0, 3, 6 };
+    int[] sample2 = new int[] { 1, 3, 2 };
+    int[] sample3 = new int[] { 2, 1, 3 };
+    int[] sample4 = new int[] { 1, 2, 3 };
+    int[] sample5 = new int[] { 2, 3, 1 };
+    int[] sample6 = new int[] { 3, 2, 1 };
+    int[] sample7 = new int[] { 3, 1, 2 };
+    int[] puzzle = new int[] { 9, 6, 0, 10, 18, 2, 1 };
 }
